// prismicio-types.d.ts
type Simplify<T> = { [KeyType in keyof T]: T[KeyType] };
// ...
type BlogPostDocumentDataSlicesSlice =
  BlogImageSlice | BlogCtaSlice | BlogTextSlice | BlogHeroSlice;

/**
 * Content for blog post documents
 */
interface BlogPostDocumentData {
  /**
   * Listing Title field in *blog post*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_post.listing_title
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  listing_title: prismic.KeyTextField;

  /**
   * Listing Excerpt field in *blog post*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_post.listing_excerpt
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  listing_excerpt: prismic.KeyTextField;

  /**
   * Thumbnail field in *blog post*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_post.thumbnail
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/image
   */
  thumbnail: prismic.ImageField<never>;

  /**
   * Category field in *blog post*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_post.category
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  category: prismic.KeyTextField;

  /**
   * Reading Time field in *blog post*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_post.reading_time
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  reading_time: prismic.KeyTextField;

  /**
   * Author Name field in *blog post*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_post.author_name
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  author_name: prismic.KeyTextField;

  /**
   * Featured field in *blog post*
   *
   * - **Field Type**: Boolean
   * - **Placeholder**: *None*
   * - **Default Value**: false
   * - **API ID Path**: blog_post.featured
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/boolean
   */
  featured: prismic.BooleanField;

  /**
   * Display Order field in *blog post*
   *
   * - **Field Type**: Number
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_post.display_order
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/number
   */
  display_order: prismic.NumberField;

  /**
   * Slice Zone field in *blog post*
   *
   * - **Field Type**: Slice Zone
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_post.slices[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/slices
   */
  slices: prismic.SliceZone<BlogPostDocumentDataSlicesSlice>; /**
   * Meta Title field in *blog post*
   *
   * - **Field Type**: Text
   * - **Placeholder**: A title of the page used for social media and search engines
   * - **API ID Path**: blog_post.meta_title
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  meta_title: prismic.KeyTextField;

  /**
   * Meta Description field in *blog post*
   *
   * - **Field Type**: Text
   * - **Placeholder**: A brief summary of the page
   * - **API ID Path**: blog_post.meta_description
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  meta_description: prismic.KeyTextField;

  /**
   * Meta Image field in *blog post*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_post.meta_image
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/fields/image
   */
  meta_image: prismic.ImageField<never>;
}

/**
 * blog post document from Prismic
 *
 * - **API ID**: `blog_post`
 * - **Repeatable**: `true`
 * - **Documentation**: https://prismic.io/docs/content-modeling
 *
 * @typeParam Lang - Language API ID of the document.
 */
export type BlogPostDocument<Lang extends string = string> =
  prismic.PrismicDocumentWithUID<
    Simplify<BlogPostDocumentData>,
    "blog_post",
    Lang
  >;